ref: make precompile errors shared
